% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord}
\alias{coord}
\title{Caluculation of the relative coordination across chromosomes}
\usage{
coord(chr, pos)
}
\arguments{
\item{chr}{Chromosome number. It should have the same length as \code{pos} and \code{p}.}

\item{pos}{Positions on the chromosomes.}
}
\value{
A relative chromosomal positions from 0 to 1
}
\description{
Coordination for the x-axis of GWAS Manhattan plots.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
