% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-point_plot.R
\name{line.point_plot}
\alias{line.point_plot}
\title{PLot at least one Y vs single X}
\usage{
line.point_plot(Input.dataset, X.data = 1, Y.line.data = NULL,
  Y.line.data.scale = NULL, Y.point.data = NULL,
  Y.point.data.scale = NULL, X.Time = F, X.limits = NULL,
  Y.limits = NULL, X.breaks = waiver(), Y.breaks = waiver(),
  X.labels = waiver(), X.minor.breaks = waiver(),
  Plot.Title = "plot Title", X.axis.title = "X axis",
  X.label.angle = 0, Y.axis.title = "Y axis", Y.label.angle = 0,
  Y.pointColor.leg = "color", Y.pointColor = NULL,
  Y.pointShape.leg = "shape", Y.pointShape = 20, Y.pointSize = 1,
  Y.lineColor.leg = "color", Y.lineColor = NULL,
  Y.lineType.leg = "type", Y.lineType = "solid", Y.lineSize = 1,
  AddLegend = F, Save.pdf = F, Save.png = F,
  Out.file.Name = "plot")
}
\arguments{
\item{Input.dataset}{A data frame with one column as X and other(s) for Y}

\item{X.data}{A column of the Input.dataset to be used as x}

\item{Y.line.data}{One o many columns of the Input.dataset to be used as y for line plots}

\item{Y.line.data.scale}{A vector of scaling factors (if needed) for Y.line.data}

\item{Y.point.data}{One o many columns of the Input.dataset to be used as y for point plots}

\item{Y.point.data.scale}{A vector of scaling factors (if needed) for Y.point.data}

\item{X.Time}{If TRUE then X is considered as date-time with corresponding x axis}

\item{X.limits}{Set limits for the x axis (if needed)}

\item{Y.limits}{Set limits for the y axis (if needed)}

\item{X.breaks}{Set of breaks for the x axis (if needed)}

\item{Y.breaks}{Set of breaks for the y axis (if needed)}

\item{X.labels}{Labels for the X axis (if needed)}

\item{X.minor.breaks}{Set of minor breaks for the x axis (if needed)}

\item{Plot.Title}{The title of the plot}

\item{X.axis.title}{The title for the x axis}

\item{X.label.angle}{The angle for the x axis' labels}

\item{Y.axis.title}{The title for the y axis}

\item{Y.label.angle}{The angle for the y axis' labels}

\item{Y.pointColor.leg}{Descriptions for the color legend for the point plots}

\item{Y.pointColor}{List of colors for points}

\item{Y.pointShape.leg}{Descriptions for the shape legend for the point plots}

\item{Y.pointShape}{List of shapes for points}

\item{Y.pointSize}{List of sizes for points}

\item{Y.lineColor.leg}{Descriptions for the color legend for the line plots.
NB: if the lists in Y.lineColor.leg and Y.pointColor.leg are the same, only one color legend
is generated}

\item{Y.lineColor}{List of colors for lines}

\item{Y.lineType.leg}{Descriptions for the line type legend for the line plots}

\item{Y.lineType}{List of line types}

\item{Y.lineSize}{List of sizes for lines}

\item{AddLegend}{if TRUE the legend is added}

\item{Save.pdf}{if TRUE the plot is saved to pdf file}

\item{Save.png}{if TRUE the plot is saved to png file}

\item{Out.file.Name}{Output file name}
}
\value{
Either a plot in 'Plots' window or a pdf/png file with plot
}
\description{
The function makes plots (line and/or points)
of at least one Y value vs a single X value. There are a single x and single y axes.
X and Y are columns from a signle data set. Type/shape, color and size of the
lines/points are controled by the function's input parameters.
Legend names for colors and shape/type are controled by the
function's input parameters. Legend position can't be controlled
by the function's input parameters. The plot can be either save
as pdf and/or png file or plotted in the Plots window of RStudio.
}
\author{
Anna Morozova
}
