% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_biPlot.R
\name{my_biPlot}
\alias{my_biPlot}
\title{Depicting a simple biplot}
\usage{
my_biPlot(x, y = NULL, data = NULL, ...)
}
\arguments{
\item{x}{A numeric vector for the x-axis. Factors are now allowed.}

\item{y}{A numeric vector for the y-axis. If NULL, the function operates like graphics::plot().}

\item{data}{A data.frame to be specified for a plot.}

\item{...}{Other arguments passed to geom_point().}
}
\description{
A simplfied ggplot function to depict a biplot.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
