% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ManhattanPlot.R
\name{my_ManhattanPlot}
\alias{my_ManhattanPlot}
\title{Depicting Manhattan plots for GWAS}
\usage{
my_ManhattanPlot(chr, pos, p, ...)
}
\arguments{
\item{chr}{Chromosome number. It should have the same length as \code{pos} and \code{p}.}

\item{pos}{Positions on the chromosomes.}

\item{p}{A vector of p-values ranging from 0 to 1.}

\item{...}{Other arguments passed to geom_point().}
}
\description{
A simple ggplot function to draw Manhattan plot for GWAS results.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
