% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_Plot.R
\name{my_Plot}
\alias{my_Plot}
\title{Depicting a simple plot}
\usage{
my_Plot(x, y = NULL, data = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A numeric vector for the x-axis. Violin plots will be shown if x is a factor.}

\item{y}{A numeric vector for the y-axis. If NULL, the function operates like graphics::plot().}

\item{data}{A data.frame to be specified for a plot.}

\item{xlab}{Option to change the label of x-axis.}

\item{ylab}{Option to change the label of y-axis.}

\item{...}{Other arguments passed to geom_point().}
}
\description{
A simplfied ggplot to depict a biplot (for continous x) or violin plot (for factorial x).
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
