% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_qqPlot.R
\name{my_qqPlot}
\alias{my_qqPlot}
\title{Depicting QQ-plot for GWAS}
\usage{
my_qqPlot(p, ...)
}
\arguments{
\item{p}{A vector of p-values ranging from 0 to 1.}

\item{...}{Other arguments passed to geom_point().}
}
\description{
A simple ggplot function to draw a QQ-plot for GWAS results.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
