% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_violinPlot.R
\name{my_violinPlot}
\alias{my_violinPlot}
\title{Depicting violin jitter plots}
\usage{
my_violinPlot(x, y, data = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{A numeric vector for the x-axis, factorized within the function.}

\item{y}{A numeric vector for the y-axis.}

\item{data}{A data.frame to be specified for a plot.}

\item{xlab}{Option to change the label of x-axis.}

\item{ylab}{Option to change the label of y-axis.}

\item{...}{Other arguments passed to geom_jitter().}
}
\description{
A simplfied ggplot to project violins and boxes on jitter plots.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
