% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{qqPlot}
\alias{qqPlot}
\title{Depicting QQ-plot for GWAS without ggplot2}
\usage{
qqPlot(p, ...)
}
\arguments{
\item{p}{A vector of p-values ranging from 0 to 1.}

\item{...}{Other arguments passed to plot().}
}
\description{
A simple plot function to draw a QQ-plot for GWAS results.
}
\author{
Yasuhiro Sato (\email{sato.yasuhiro.36c@kyoto-u.jp})
}
